jwt_auth_project/requirements.txt
asgiref==3.8.1
Django==4.2
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.0
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
PyJWT==1.7.1
pytest==8.3.5
sqlparse==0.5.3
